refactor(P2.6 slice 3): move the WITH-discovery cluster into with_to_cte
Third sub-slice of P2.6 (REFACTORING_SAFETY_PLAN.md §5.1). Moves the contiguous
WITH-discovery / join-pruning cluster (old lines 10803–11753, ~951 lines)
verbatim from plan_builder_utils.rs into render_plan/with_to_cte/mod.rs:
`find_all_with_clauses_grouped`, `collapse_passthrough_with`,
`node_is_concrete_labeled`, `alias_has_pattern_correlation`,
`prune_joins_covered_by_cte`.

Byte-identical: the moved block diffs clean vs the original modulo one widening.
`alias_has_pattern_correlation` widened `fn` → `pub(crate) fn` (documented) —
it turns out `build_chained_with_match_cte_plan` calls it directly (two sites),
so it is re-exported alongside the three already-`pub(crate)` names. Only
`node_is_concrete_labeled` stays private (its sole caller,
`alias_has_pattern_correlation`, moved with it). The head `#[cfg(test)] mod
tests` (still in plan_builder_utils) resolves its 9 `find_all_with_clauses_grouped`
calls via the re-export + `use super::*` (P2.4 precedent).

Cross-module deps now resolve via targeted imports in with_to_cte:
`find_cte_reference_alias` (alias_utils, was reached via the `alias_utils::*`
glob) + `plan_contains_with_clause` (plan_predicates, already present). The
slice-2 transition import of `find_all_with_clauses_grouped` from
plan_builder_utils is dropped (now defined in this module).

Ratchet net-zero (the cluster carries no schema-axis tokens — no baseline
change). Gates: 1612 lib + 529 integration (incl corpus_sweep + sql_golden) +
1 ratchet + 7 unit, 0 failures; goldens + corpus byte-identical (no churn);
fmt/clippy clean; warning-free --tests build. Next: P2.6 slice 4 (the
build_chained giant + the 24-util widening + finalize).
